+++ RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v0.2.4-alpha.2] .
 ### Releases
 - [NPM v0.2.4-alpha.2] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.2 - .
+
+## 2025-03-11 21:16:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.0 (2.548.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.4-alpha.3] .
+### Releases
+- [NPM v0.2.4-alpha.3] https://www.npmjs.com/package/@inkeep/inkeep-analytics/v/0.2.4-alpha.3 - .